Split Filter and Specimen models into numerous files
