+++ CopeID.Models/Specimens/SpecimenGender.cs
@@ -1,0 +1,8 @@
+﻿namespace CopeID.Models.Specimens
+{
+    public enum SpecimenGender
+    {
+        Male = 0,
+        Female = 1
+    }
+}
